backend/requirements.txt

aiosignal==1.2.0
appnope==0.1.2
APScheduler==3.8.1
argh==0.26.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
backcall==0.2.0
backports.functools-lru-cache==1.6.4
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bleach==4.1.0
blinker==1.4
brotlipy==0.7.0
cachetools==4.2.4
certifi==2022.6.15
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
coverage==6.3.1
cryptography==36.0.1
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.2.0
email-validator==1.1.3
emoji==1.7.0
entrypoints==0.4
executing==0.8.3
filelock==3.6.0
Flask==2.0.2
Flask-Cors==3.0.10
flask-mongoengine==1.0.0
Flask-WTF==1.0.0
flit_core==3.7.1
fonttools==4.29.1
frozenlist==1.2.0
future==0.18.2
google-auth==2.3.3
huggingface-hub==0.4.0
idna==3.3
importlib-metadata==4.11.2
importlib-resources==5.4.0
iniconfig==1.1.1
ipykernel==6.9.1
ipython==8.1.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.1.0
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
json5==0.9.5
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyterlab==2.1.0
jupyterlab-pygments==0.1.2
jupyterlab-server>=1.1.0
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
lxml==4.8.0
Markdown==3.3.6
MarkupSafe==2.1.0
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
mongoengine==0.24.1
multidict==5.2.0
munkres==1.1.4
nbclient==0.5.12
nbconvert==6.4.2
nbformat==5.1.3
nest-asyncio==1.5.4
nltk==3.6.7
notebook==6.4.8
numpy==1.22.3
oauthlib==3.2.0
opencv-python==4.5.4.60
packaging==21.3
pandas==1.3.5
pandocfilters>s=1.5.0
parso>=0.8.3
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pip==22.1.2
pluggy==1.0.0
prometheus-client==0.13.1
prompt-toolkit==3.0.27
protobuf==3.19.1
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.8.1
pycodestyle==2.8.0
pycparser==2.21
Pygments==2.11.2
PyJWT==2.3.0
pymongo==3.12.3
pyOpenSSL==22.0.0
pyparsing==3.0.7
pyrsistent==0.18.1
PySocks==1.7.1
pytest==7.0.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==22.3.0
regex==2022.3.2
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
sacremoses==0.0.46
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
Send2Trash==1.8.0
setuptools==59.8.0
six==1.16.0
smmap==5.0.0
soupsieve==2.3.1
stack-data==0.2.0
statsmodels==0.13.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
terminado==0.13.3
testpath==0.6.0
threadpoolctl==3.1.0
tokenizers==0.12.1
toml==0.10.2
tomli==2.0.1
torch>=1.10.0
torchvision==0.11.2
tornado==6.1
tqdm==4.63.0
traitlets==5.1.1
transformers==4.19.0
tweepy==4.4.0
typing_extensions==4.1.1
tzdata==2021.5
tzlocal==4.1
unicodedata2==14.0.0
urllib3==1.26.8
waitress==2.0.0
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.3
wheel==0.37.1
widgetsnbextension==3.5.2
WTForms==3.0.1
yarl==1.7.2
zipp==3.7.0
